Ensure Windows installer exits on attempt to downgrade
Generally I think this should be allowed, however when code was refactored in #6396 this was missed out and the installer doesn't exit. Since it also doesn't correct manage the Windows services on this flow, it doesn't work correctly if you continue and you get a whole lot of errors. If you skip these errors, and manually interfere with services you end up with a weird install with the downgraded code but with Windows thinking the new version is still installed. Probably better not to allow this.