Update material userName/password in sync with one another
When loading a triggered run from the DB to construct buildWork, the userName from the material is, for some reason inflated/persisted. However the password is not. So when determining the correct password to use for construction of `BuildWork` to dispatch to the agent, it will use the one from the `MaterialConfig` attached to the pipeline, not the de-duplicated material used for triggering.

If the pipeline is configured with a different username for the same url/branch in two different pipelines, the password will be updated but not the userName, leaving them out of sync. For some targets this doesn't matter when using tokens (e.g GitHub) but others require them to match, including those that require legacy username/password. In any case, the entire credential needs to be updated together.