Normalize client-sent `If-Match` and `If-None-Match` headers to strong Etags
Currently GoCD always sets strong ETags in response headers as noted in #6278. Putting aside whether this is wise or not (as opposed to setting weak etags), we use these etags both to determine whether a resource is modified or whether there has been a race condition update in `If-None-Match` and `If-Match` headers respectively. https://datatracker.ietf.org/doc/html/rfc7232#section-3.1 says that when comparing to the If-Match header we should use "strong comparison" which effectively means the a weak etag validator is not considered equal to the strong etag validator with the same content. Unfortunately this is a problem because some caching proxies (e.g CloudFlare) will minify JSON responses for optimization and convert our strong etag to a weak etag. This causes updates to fail.

Technically we already compare etags in a slightly "non-strong" way because the server ignores `--gzip` and `--deflate` suffixes that Jetty can add when compressing responses. It also doesn't really make sense to compare objects "byte-by-byte" when considering API responses which do not really have a canonical octal/binary representation.

In any case, it seems to express the content more clearly to have our UI JS normalize the ETags it sends back to strong ETags. Alternative would be to change the matching on the server side to allow a weak etag to match a strong etag, but this seems to go against the spirit of the spec. However the downside of this approach is that use of the GoCD APIs via other client code will likely have similar issues, and have to normalize their own ETags in PUT calls made to the API.