Change GoCD to no longer force links to open in new windows/tabs
Internal links will open by default in the same window/tab, returning control to the user to be able to middle-click for new tab/window, and shift/middle-click for new background tabs, reducing a relatively age-old user complaint about excessive tabs being opened, and getting lost in tabs.

External and docs links continue to generally open in new tabs for now; with this change to make the "external link icons" consistent across places, so you have a good
idea if something is going to open in a new tab from the visual description of the link. There are some old non-Typescript pages which don't do this still unfortunately.