Workaround OpenJDK 17.0.3 issue breaking parts of Rails/Rack on Windows
Apply same workaround as commit 98e395ab2cdbcd109d17a76b97fc23d7675edde5 to the server. On Windows Rails/Rack seems to fail to initialize due to a bundler problem (?) due to use of the NUL device without this. We should remove these workarounds once we upgrade to JDK 17.0.4+. Note that Oracle JDK released a patch in 17.0.3.1 for which Adoptium/Eclipse Temurin did not release a similar patch release.

Flag appears to only affect Windows