Correct wrapper log location for docker images
Seemed to be a typo in the sed command that updates the wrapper log location during startup. Intent is surely to store the logs in /godata/logs so they persist after restart, rather than using the root working dir. Currently it's not working correctly, thus storing the log at `/go-working-dir/wrapper.log` as default location.