Remove unused XhrPromise
The last use of XhrPromise itself seems to have been removed in 169bd08. Unfortunately, this included an intended polyfill for Promise.finally on IE11 which we shouldn't need anymore. Unfortunately due to a logic error (in the FIXME), the Polyfill has been overriding the browser `Promise.finally` all the time, rather than acting as a polyfill. Removal of the "polyfill" causes some autocompletion tests to fail, so needs more analysis to understand whether it can be safely removed or some real logic is relying on this different Promise.finally behaviour from browser. This can be done separately.