Fix Chrome/Chromium navigation on Stage graphs/stats iframe sandbox (older/newer links)
With the iframe sandbox, on Chrome the authentication cookie is not sent for clicks on the links "inside" the iframe. (it's OK when the iframe is loaded from the parent page) This causes the server to think you are logged out, so you see a blank iframe for the graph. This is because Chrome treats Cookies without a SameSite attribute as Lax, and won't send them via the iframe which is currently considered "cross-origin" due to the way the sandbox is defined. Other browsers are fine, because they assume SameSite=None when it is not specified, a less secure option.

We could set sandbox="allow-scripts allow-same-origin" however this leads to warnings on same browsers, as the combination would technically allow the iframe to remove its own sandbox, so gives you a false sense of security. We could also remove the sandbox entirely. However an alternative is to use the same `postMessage` mechanism we are currently using to allow the iframe to change the parent window location. We can use this to also allow it to change its own location and send the cookie correctly, since the action is triggered from the parent context.