Remove unnecessary iframesandbox flag for job details tests tab
This has been defaulted to true for a very long time. If it's causing problems we are probably better to just remove it. (I can imagine there might be issues with clicking any links within the template, but that doesn't seem to be possible at the moment in the generated page, so perhaps not a concern.