Improve handling when siteUrls are configured with whitespace only
Currently this causes confusion because they are correctly validated by the UI, but then cause authorization plugins to 500 when determining the root URL to use. This is caused by different validation logic vs interpretation logic. This change aligns the logic used by plugins to be based on blank = "not set", consistent with the validation.