Exclude assets pipelines Gems from production package
In earlier Rails world, it seems there was a dedicated `assets` group, but this was removed as it caused some problems. This re-introduces the concept here in order to be able to run without these Gems during production. Don't follow understand how this all works, but

- GoCD doesn't do a secondary `bundle install` after compiling assets. It just uses some random code to exclude non-default group Gems from the production war file. So from bundler's perspective they are still there.
- Thus we need to override during `boot.rb` to tell it to only setup the `default` gems, not everything implied by the `Gemfile` and `.bundle/config`, which has no `BUNDLE_WITHOUT` set. I use `RAILS_ENV=production` and `!RAILS_GROUPS` to determine this
- During asset compilation, tests, local dev `RAILS_GROUPS` is set to tell Rails to "add in" the assets pipelines gems to be able to compile on the fly.
- During `init.rb` some extra magic is required to require the rails groups, that I dont fully understand.
- We need to disable the `jsroutes.rb` initializer when running in production. Not sure if this should be an initializer but this seems to work fine.

These ideas came from https://nts.strzibny.name/removing-nodejs-dependency-from-a-rails-application/ and https://stackoverflow.com/a/39758749 along with a lot of trial and error. Some of the complications come from the JRuby-Rack based bootstrapping process, which seemingly works differently to other regular Rails apps in terms of env var visibility as well as recognition/ordering of `boot.rb`, `init.rb`, `application.rb` etc.