RUBY env var needs to be forward-slashed even on Windows
Without it being forward slashed, native builds don't seem to be possible and the back-slashes get stripped. Strange, but true. We don't currently rely on this on Windows, it seems, but it took a while to fix in order to unblock an experimental native compile of libsass/SassC for Windows so seems sensible to integrate.