Upgrade from Sprockets 3 to Sprockets 4
Sprockets 4 introduces need for mandatory app/assets/config/manifest.js (which is not actually JS) that replaces the old assets.precompile block. By default everything in the root of the load paths (app/assets/stylesheets etc) seems to be treated as an entry point which is not what we want, so continue to list out the individual entry points we need.

https://github.com/rails/sprockets/blob/main/UPGRADING.md

This
- removes a whole lot of dead entrypoints for stylesheets migrated properly to webpack land
- changes image loading to not just include ANY image in the load path (it's not possible like it was before anyway, it seems) and only include those referenced. This doesn't seem to have any effect.
- only includes woff/woff2 versions of fonts based on their inclusion. The others aren't required anyway it seems?