Fix rendering of elasticAgent job detail headers
Previous code assumed these were client-side vars, which was wrong. They are server side rendered.

Even after this fix, this code is a bit buggy, but this restores it to previous Velocity-based state. Thus bug is that at time of server-side rendering the `elasticAgentId` may not be known even though it *is* known that the build will run on an elastic agent. The Error is that the `{if}` clause is client side, and changes over time, but the client side is never informed of the `elasticAgentId` when allocated, and so it is unable to change the status report link to the specific elastic agent once it is "known". This change doesn't fix the race condition to avoid needed to refresh the entire page, but does fix the WTF in the code.