Reset autosuggestions every time the pipeline config route changes
Auto-suggestions are based on upstream stages and pipelines from the current pipeline, so in a multi-stage pipeline, the list of upstreams change when you switch stages. This means even resetting on save isn't sufficient, you really need to reload suggestions every time the context for the tab changes. Since the content object is re-used within a single pipeline's config, we need to detect this and reset when the content is changed to avoid giving the user "stale" auto-suggestions that don't reflect a new stage/job added, or ones intended for a different stage.