Improve teardown ordering
It seems that the ordering between the @AfterAll and the @TempDir cleanup might not be deterministic here, and sometimes causing attempts to remove the TempDir before the Perforce daemon stops, which causes a failure on Windows due to locks. Try using a static field rather than injecting to @BeforeAll to see whether that helps make the ordering more deterministic?