Allow Git submodule tests to pass after fix to CVE-2022-39253
By default git submodules cannot be references to local file repos after these fixes. This adds an opt-in to set the appropriate config option for each operation, used only by tests by default. Theoretically available for users to override if necessary, but I don't see why use of file reference submodules is a legitimate use case in a CI system anyway, so don't plan to publicise it for.