Ensure we upgrade dependencies for all images
Currently we do so for alpine/dind, but not debian/ubuntu/centos so are relying on the
upstream docker image to regularly be updated. This does tend to happen regular for
most images as they are official Docker images (except CentOS), but doing so ourselves
allows us to ensure we get patches faster and with confidence.