Remove legacy Windows quoting strategy deprecated in 22.2.0
Since August 2022 release there have been no issues reported with the change in quoting approach, so removing this old feature
flag which is likely no longer required. Feature flag wasn't published and only kept as a backup in case of unintended
regressions, so unlikely folks are using it, and have interacted with some folks using Windows more extensively recently so
have a reasonable degree of confidence we have this correct.

Will keep the nested tests here in case we want to go back and add proper coverage for POSIX/non-Windows setups.