Shutdown goCaches properly within Spring lifecycle
Currently tests have confusing cascading failures when spring start-up fails, because EhCache caches are not de-registered from the
static CacheManager, causing subsequent tests to fail in a very misleading way due to a duplicate cache registration. Removing the
cache during the Spring pre-destroy lifecycle is a better way to manage this. Seems to be fine with a real running server but this
is possibly because we are not relying upon the Spring context shutdown lifecycle properly, so we need to monitor this.