Crop overly long log lines
Log lines exceeding a configurable threshold are cropped on agent
side and a "...[ cropped by GoCD ]" notice appended to the line.
This prevents extremely long log output from crashing the server with an
Out Of Memory error.

The threshold defaults to 1 million characters per line, and can be
configured on the server side like so:

-Dgocd.agent.extra.properties="go.console.log.max.line.length=1000000"

Cropping can be disabled by setting the threshold to 0.

This solution may still crash an agent, since a line is first read fully
into memory before cropping. Rationale:

1. Protecting the central server is more important than individual agents
2. Secret redaction requires that the full line is known, otherwise we
  risk leaking a partially cropped secret.
3. Reading output into memory in a bounded fashion requires a deep dive
   into Java standard library code (i.e. BufferedReader). This could
   cause additional effort to keep in sync with later improvements or
   fixes (though it probably changes rarely).

The last two points are not impossible to solve, but the cost/benefit
is questionable.

The cropping is implemented as an additional OutputStreamConsumer decorator
and applied only in CommandLine::runScript. This method is used for
user-defined pipeline tasks, but not for other commands, such as SCM
tasks, where cropping is undesirable.

Fixes #11100