Remove gcc-libs from Alpine images
gcc-libs is large, and provides libs that don't seem to be required for general JRE usage, or by GoCD in particular, based on https://archlinux.org/packages/core/x86_64/gcc-libs/

These were possibly here as dependencies for JDK tools rather than the main JRE, or because the AdoptOpenJDK image this originally came from was more general purpose. Removing these makes it a little easier to evaluate creating an aarch64 image.