Update maintenance mode API to shared v11
Only used to represent which materials have running updates, so will validate this is fine. Strangely api-shared-v1 seemed to be introduced at a time when there was already an api-shared-v4 and api-shared-v6 (decreasing the version number) it's it's not entirely clear why.

Nevertheless, the maintenance mode API seems to work the same with shared v11 including via GoCD UI, and even minor changes in the material representations shouldn't matter as these are not explicitly documented via api.gocd.org and only affect the "Running MDU" response portion.