Remove old monkey-matching of Ajax
This should not be needed now, and is not done consistently anyway. Was some kind of IE hack to deal
with caching. Let's move forward and assume things work properly based on headers set by the server.

There don't seem to be many legacy JS/AJAX places reliant on this right now anyway.