Switch main agent work loop to an exponential backup impl
Prior to this change, agents had an initial once-off delay (1s) and then only
attempted to get work every 10s. During start-up, and transitioning between pieces
of work this caused a lot of unnecessary delay, as it takes two server attempts to get
a new piece of work - the first where the server doesn't yet know the agent is idle, and
the second where it has asynchronously allocated some work.

This impl changes the agent so every time it completes a piece of real work, it resets the
wait time to the "initial" delay, and then does exponential backoff of attempts up until the
configured interval. This has the effect of making agents more responsive to picking up work
when they initially start, and when transitioning between one piece of work and the next, at
the cost of some more calls to the server.