Make wrapper flush logs line-by-line by default
The current configuration is buffered, which means you get half-logged lines which are extremely confusing, and not really the norm
these days. Don't expect any major performance impact from this, as the content logged is just not that huge, and in most cases
not piped to the wrapper on the console.