Close modals only on "cancel-click" or "ESC"
Some modals on the pipeline config SPA close when you click on the overlay background, but some do not.
This makes them consistent, which also mitigates browser inconsistent when you drag to select text and the MouseUp
event is "outside" the overlay. To figure out the intended "target" you need to know target of both MouseUp and MouseDown
which is a pain. Better to disable this on modals that have text areas folks might be copying+pasting to.