Disable redirect following on agent upgrade checks
This is a bit of a safer default in the event of server-side authentication/authorization issues, especially re-authentication. Prior to this change,
and fixes to the endpoint, with re-authentications enabled, the server can erroneously redirect the agent status API to the login page. By default
the client will follow redirects and get a response that is not the API response it is expecting. This causes NullPointerExceptions looking for the
appropriate response headers. Disabling redirect following doesn't change the end result, but avoids confusing NullPointerExceptions, changing them
to ERROR logs due to the `302` response that is unexpected.