Fudge to get most tests to pass on JQuery 3
JQuery 3 makes .then() in Deferreds async to fit with ES6 Promise spec even if the promise is pre-resolved
as happens when mocking AJAX with mock-ajax. Unfortunately this breaks many tests because they are relying
on the ajax promises resolving synchronously before doing other things. Some code can be changed to async/await
style to allow waiting on promises in the tests before proceeding, but other stuff is deep inside code and a
pain to correct (e.g the pipeline/dashbboard widget specs). In these cases the easiest thing for now is to
add dodgy fudge-factor sleeps.

In many cases we also need to force a mithril redraw for the same reasons, if the expectations afterwards
depend on a redraw. In some senses these redraw cases may actually be exposing a bug in the production code
- perhaps there should be a redraw() in a .always() after the promise chain is resolved as there is in other
cases, so just delaying until the promise is resolved is sufficient on its own.