Run frontend lints alongside jasmine task
Currently this runs duplicated on the divide-and-conquered agent/server unit tests
which is confusing as multiple jobs fail. It makes more sense to run this alongside the
jasmine tests since they run on JS for both old and new JS, they all depend on yarn install
and they now run at the start of the pipeline anyway, for fast feedback.