Workaround issues with Alpine 3.18 on docker-dind image
When muslc-compiled libudev is installed on Alpine 3.18 it seems to cause a JVM crash when the OSHI library tries to determine OperatingSystem details, because it tries to load the udev lib. Somehow this worked OK on Alpine 3.17 but fails on Alpine 3.18 despite the version of udev seemingly being the same.

libudev via the eudev-libs package is only required by the btrfs storage driver. I think we can get away without this for now, so removing it from the built image.