Fix mass duplicate imports of font-awesome CSS
This is causing huge bloating of all sprockets asset pipelined CSS, because font-awesome-glyphs (which should really go away)
is being included in many places. Since this was also duplicating an inclusion of font-awesome and bourbon the CSS bloats massively
especially with the much larger number of icons in Font Awesome 6. We should rely on these imports having been done elsewhere if
the user wants to use font-awesome, this file is just some mix-ins to make glyphs easier to use.