Accept auth session context from Authorization Plugins
As per the linked issue this commit accepts some additional context (e.g. a state param and nonce) from an Authorization plugin during the `getAuthorizationServerUrl` flow.

This context is then added to the users session so that it can subsequently be used by the plugin for verification etc. This will be added in a later commit.

https://github.com/gocd/gocd/issues/11629