Correct Auto Update toggle for SCM materials to be disabled when material is readOnly
Confusingly there seems to be a concept of both disabled and readonly for material fields, and not sure
how they are respectively used. Rather than rationalise all this, identify that when defined in a config
repo it is "readonly" that seems to be true, and all inputs except for the material AutoUpdate toggle
are already correctly disabled. This corrects the toggle.