Don't completely overwrite the agent wrapper-properties.conf within container entrypoint
Custom entrypoint scripts run before the main entrypoint, and should be allowed to add things to the pre-packaged wrapper-properties.conf. Users
will need to ensure the values don't overlap if they do so, however they should be allowed to use their entrypoint to do this, as works on the
server image entrypoint.