Upgrade and consolidate transitive dependencies for "opinionated" resolutions
Can mainly do this with yarn up -R '*' now, however it does remove the force resolutions for vulnerable old dependencies. May need to resolve this in the proper yarn 3 way. For the record the old crude Yarn 1 forced resolutions were the below, which caused some downgrades in some cases, as well as some unnecessary upgrades due to limitations of the syntax.

  "resolutions": {
    "**/chokidar/**/glob-parent": ">=6.0.2",
    "html-webpack-plugin/**/ansi-regex": "^5.0.1",
    "**/loader-utils": "^2.0.3",
    "**/semver": "^7.5.4",
    "**/string-width": "^4.2.3",
    "**/wrap-ansi": "^7.0.0"
  }