Ensure use of Spring TypeConverters is thread safe, and re-used
At some point TypeConverters were decided they should be re-used when loading config from XML; perhaps there was too much garbage
being generated. In any case, these are not threadsafe. Perhaps there was only a single thread trying to use them at the time, but
it certainly seems possible for multiple to try and do it now, and bizarre behaviour happens when doing so, such as not being able to
convert a string 'false' to a primitive boolean and even returning null, which won't work.

Additionally, TypeConverters used when writing XML were NOT being re-used which seemed inconsistent. I don't know if they really need to
be re-used/cached but probably better to do so consistently. Only risks here are that we have too many cached across threads (wasteful)
or alternatively that we have ephemeral threads being created for these purposes, thus making the caching ineffectual, but I don't think
either are true. If this turns out to be the case, could use a caching object pool or some other mechanism.