Ensure Dates/Timestamps are serialized/deserialized deterministically
Default GSON serialization has some nondeterminism due to use of default locales and parsers. While this
generally should work fine, there are changes made to these default formats in Java 20 and use of NBSP vs
space which breaks things. Generally speaking it'd be better to use a standard serialization such as ISO instants
with milliseconds anyway, to represent java.util.Date objects.

Ideally we shouldn't need to send Dates to agents anyway, but this needs a separate look to determine why modifiedDate is being
sent across and whether this is appropriate.