Ensure webpack builds are clean when changing node_modules
In some cases the webpack cache at rails/tmp seems to contain stale files that need to be rebuilt
after a yarn install of new packages, which is currently not being done in some cases. Webpack 4's
cache-loader doesn't seem to detect that node_modules etc has changed and it should invalidate the
cache so I think we need to do it for it to avoid inconsistent local builds compared to a clean environment.

For now, if we decide to recompile assets via Gradle, we will do it without cache. Webpack in watch mode can
still rely on the cache. Another alternative is to make the cache clean conditional on yarn install having changes
but that might not be enough (e.g if .babelrc or tsconfig change are not detected) so let's just ensure things are clean.