Correct test expectation for Jackson 2.16.1
Jackson 2.16+ redacts source from exception messages by default. I don't think this
matters for GoCD at runtime, but it does cause this test assertion to fail, as it's looking for specific text
to make sure the logic to force some bad JSON in `JsonOutputWriter` is working. Don't really understand the
background of why the JsonOutputWriter behaves this way, however probably better to err on the side of not leaking
data in the logs for now, so not changing the default back for the JsonOutputWriter to include source.

See https://github.com/FasterXML/jackson/wiki/Jackson-Release-2.16#streaming-jackson-core and https://github.com/FasterXML/jackson-core/issues/991