Improve performance of agents API responses
Avoid putting EnvironmentConfigs into HashSets - computing hashes and equality
on this is extremely expensive, iterating all pipelines etc. It's also not necessary
for the purpose of these APIs; a list is perfectly fine and uniqueness is ensured elsewhere.
Furthermore, for the massive index API, a single Map can be built in one pass, rather than building
many Lists.