Improve efficiency and clarity of working with ServerHealthStates
Previous code is especially inefficient, and when you have a large number of errors, spends a lot of time in collection processing
logic trying to remove errors (due to creating new temporary collections over and over). This uses the collections a little more efficiently,
although there is a lot of inefficient code still around here, sorting unnecessarily and otherwise distributed around the codebase.