Fix toggle of show/hide password on modal pages
On modals with passwords when the review re-renders the password is auto-hidden as the state is forgotten due
to previous implementation hacking with CSS attributes directly. Believe the right way to fix this is
to record the state within the control and render off this rather than hacking styles directly.

Also fixes the icon display (icon should show next action, not current state) and wrapping of reset/change button to next line.