Fail Linux package installs if a compatible JRE cannot be found
Not entirely sure about this change, as it forces the JRE location to be correct (and able to be found by apt/dpkg/dnf) before installing GoCD itself, however
this helps deal with cases where the recommended package dependencies are not installed by default on a user's system and they don't notice the warnings.
In the much rarer case where the user intends to override the Java location using wrapper-properties.conf they'd need to pre-create that file without
having a template/example, which is not ideal.

It also has some weaknesses in not respecting PATH overrides inside wrapper-properties.conf, not respecting go user specific configuration, not
validating the correct Java version etc - but the aim is to be better, not perfect.

Let's see how this goes, and revert if it is problematic.