Switch from curl image to Chainguard bash images as utility image
The curl image currently has issues on latest buildkit due to being built with an old Buildah version which includes
xattrs in tar layers which Docker 25, and especially rootless DIND/fuse-overlayfs (which we use for builds) do not like.
Switching to one of the Chainguard images is actually better. Images are similarish size and rebuilt more regularly.