Remove copyright license year date requirements
This does not appear to be strictly necessary, and a number of projects have removed the need for such pointless maintenance for source code, including curl and other projects. https://daniel.haxx.se/blog/2023/01/08/copyright-without-years/ It also appears to be dubious to update the date every year without any substantial revision to the work, so it's probably better to
just let source control do the talking here.