Fix display of stage overview models for VSM nodes which have periods in their IDs
There is some historical weird escaping of node IDs for the VSM view which needs to be matched by
login in the shim to display modals, or the node cannot be located correctly.

It's difficult to reuse code between the legacy JS and the typescript SPAs here, so duplicating the logic for now.

The old CSS.escape should not be needed, and looking back I cannot understand why I added it.