Avoid multiple stage overviews being displayed on VSM when server is slow to return model
If the server is slow to return it's possible to lose track of modals that need closing and the UI will get
stuck with multiple rendered. This is due to the state.model() check which only succeeds when the server has returned,
even if the user has changed their mind and clicked a different stage.

This check was being used as a workaround to avoid the `onclick` handler for clicking outside a modal firing afterwards
and then immediately closing the same modal, which can be handled better by preventing propagation for clicks inside a stage bar. Sadly due to the legacy
onclick handlers here this is being prevented with a deprecated browser property as not sure of a better way :-/