Increase UI responsiveness by decreasing default refresh interval to 5 seconds
This should be OK for most folks, although could potentially cause issues for some by doubling server requests. The current 10 seconds is a really long
time (especially when triggering UI actions like new builds), and since we seem not ready to move to websockets this seems
the best option. If they have very slow servers, folks can revert to the old behaviour with `-Dgo.spa.refresh.interval=10000`
so let's give this a go.