Remove BouncyCastle from server through DES cipher generation removal
There should be no need to ever encrypt with DES anymore; we only support decryption during migration to AES. Accordingly, we can replace the only other usage of BouncyCastle (to verify update packet RSA signing) with basic usage of default crypto libraries, since the public key is in a known format bundled with the server.

This also removes an ancient cipher updater which ran at start-up to remove an insecure old cipher and re-encrypt. This should not be required now, as the config migrated should remove all DES encryption and convert to AES anyway, whether the DES cipher is safe or not.