Move module substitution rules into dependency resolution strategy
For some reason these don't apply consistently to projects like spark-base and api-base when specified as dependency blocks so some configurations are unaffected (e.g spark-base compileClasspath and many others). Perhaps easier to try and just apply them to every configuration. Unfortunately that doesn't seem to be an option for the transitive exclusion logic, but that one is less important to apply universally, as long as it applies correctly to the runtime classpaths.