Re-integrate dependencies into build.gradle
As noted in https://github.com/dependabot/dependabot-core/issues/12559 dependabot seems to have broken for some reason. It doesn't look like it will be fixed, as has been broken for several months now.

Try to workaround this by re-integrating into build.gradle and see if we can continue the current approach. If this fails, will need to migrate to toml deps I think.