Revert use of bundle exec for rake/rspec tasks
This causes issues on Windows using jruby jars directly as addressed earlier, but forgotten, because
the bundler exec stub files for Windows don't work properly, or seem to allow override of jruby.exe
with our own binstub (the one we generate as server/scripts/jruby.bat). Sadly, need to revert to using
raw rubygems to make this work on Windows it seems.