Declare anonymous VOLUMEs for Go's working and data directories
Without these, if the user does not mount a directory, the agent uses the overlay filesystem storage drivers on top of the image, which are more complex and a lot slower than using anonymous volumes. Container orchestration platforms should clean these up automatically when the container is stopped, but some review of the docker elastic agent may be necessary to avoid volumes piling up.

This should be a net performance boost for almost all cases, since on elastic agents /godata is rarely mounted. On the server, this should be neutral except in testing cases, since folks will be mounting a /godata over the top of the container filesytem already, which will still go over the top of this anonymous volume.