Add back PIPELINEMATERIALREVISIONS.FOLDER to pipeline timeline query
While unused in code; and unimportant for the timeline, there seems to be no ordering defined for this query and the sort in code is not fully explicit. The ordering of the PMRs/revisions seemed to somehow change after this; at least on H2 implying different natural orders than persisted, causing failures like `Calculated natural ordering 1.0 is not the same as the existing naturalOrder 1.5, for pipeline Security, with id 13379`

Try adding it back to the query to see if it resolves things.