Correct stability of PipelineTimelineEntry comparison after refactoring
The factoring started to "properly" compare entries for the same pipeline entry that had the same counter value. This theoretically shouldn't happen, I believe, but there are some cases that seemed to be visible in build.gocd.org database where the pipeline name had different capitalization. In these cases the previous 25.3.0 compareTo broke the contract as A > B and B > A according to the code. In previous refactoring commit I "fixed" this accidentally by using `Integer.compareTo(a,b)`, but this caused the natural order comparisons to calculate differently in this case because the TreeSet of entries would order differently than earlier for these two entries. To resolve this; adds a third tier comparison on the pipeline ID which should safely tie break, as well as match the earlier behaviour, due to the way matches were sorted from the DB (pipeline.id, then pipelineMaterialRevision.id).