Correct stopping of service before RPM upgrade
It seems this was not working for any case except legacy direct use of /etc/init.d; so if running via systemd it was not stopping the service before upgrade. This might explain certain issues folks have had during upgrades, especially when using a CIFS mount which might prevent deletion with open file handles. This changes to rely on Tanuki to check and stop the service using its logic, similar to on Debian.